refactor: Refactor file names and organization. Refactor compute view and styling.
