+++ src/features/Compute/CaseGroup/CaseGroup.styled.tsx
@@ -1,23 +1,6 @@
 import styled from 'styled-components';
-import { spacings } from '../../../../tokens/spacings';
-import { theme } from '../../../../tokens/theme';
-
-export const CaseCard = styled.div`
-  display: flex;
-  flex-direction: row;
-
-  padding-top: ${spacings.LARGE};
-  column-gap: ${spacings.MEDIUM};
-`;
-export const Case = styled.div`
-  display: flex;
-  flex-direction: column;
-
-  padding: ${spacings.LARGE};
-  column-gap: ${spacings.MEDIUM};
-
-  border-bottom: solid thin ${theme.light.ui.background.medium};
-`;
+import { spacings } from '../../../tokens/spacings';
+import { theme } from '../../../tokens/theme';
 
 export const Parameters = styled.div`
   display: flex;
@@ -34,4 +17,10 @@ export const Wrapper = styled.div`
   border-width: ${spacings.SMALL};
   border-color: ${theme.light.primary.resting};
   border-radius: ${spacings.BORDER_ROUNDED};
+`;
+
+export const CaseList = styled.div`
+  display: flex;
+  flex-direction: column;
+  row-gap: ${spacings.MEDIUM};
 `;
