+++ src/features/Compute/CaseGroup/CaseOptionSelects/CaseCardParameters.styled.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
-import { spacings } from '../../../../../tokens/spacings';
-import { theme } from '../../../../../tokens/theme';
+import { spacings } from '../../../../tokens/spacings';
+import { theme } from '../../../../tokens/theme';
 
 export const Parameters = styled.div`
   background-color: ${theme.light.ui.background.default};
