+++ src/features/Compute/CaseGroup/CaseRow/CaseRow.Styled.tsx
@@ -1,0 +1,21 @@
+import styled from 'styled-components';
+import { spacings } from '../../../../tokens/spacings';
+import { theme } from '../../../../tokens/theme';
+
+export const Case = styled.div`
+  display: flex;
+  flex-direction: column;
+
+  padding: ${spacings.MEDIUM};
+  column-gap: ${spacings.MEDIUM};
+
+  border-bottom: solid thin ${theme.light.ui.background.medium};
+  background-color: ${theme.light.ui.background.default};
+  border-radius: ${spacings.CARD_ROUNDED};
+`;
+
+export const CaseRow = styled.div`
+  display: flex;
+  flex-direction: row;
+  column-gap: ${spacings.MEDIUM};
+`;
